Enable mypy type checking for duolingo_base.registry (#353)
## Description
### Enable type checking unit tests
The previous unit tests didn't actually do anything for the type
checking as mypy skips the body of any function that doesn't have any
type information.  All `test_` methods were of this form.  Adding a
return value type hint of `-> None:` solved this.

### Enable type checking of the `registry` sub-package
Type checking was turned off via the `setup.cfg` file for most of this
repository.  This has been re-enabled for the registry and its unit
tests.

### Enable use of inline type hints for consumers type checking their repos
Per PEP-561, if a package wants to export its inline types to package consumers (as opposed to stub
files or a typeshed `types-{package}` package) it needs to add a `py.typed`
file to the appropriate sub-package (or to its root) and make sure it is included in the built tarball.

As this package uses inline types as opposed to stub files I've added a
`py.typed` file to the `duolingo_base.registry` package. Changes to
`setup.py` allowed more sub-packages to be added by simply including
such a file in their directory.

## Verification steps

- [x] mypy breaks on bad unit test types (tested locally)
- [x] mypy passes type-checked unit tests.
- [x] Type hints work in `duolingo_base.registry` when installing built package from tarball in other repo.